minor fixes
git-svn-id: svn://svn.r-forge.r-project.org/svnroot/msgl@123 edb9625f-4e0d-4859-8d74-9fd3b1da38cb

+++ pkg/sglOptim/inst/include/rtools/rList.h
@@ -93,7 +93,7 @@ public:
 		{
 			if (name.compare(names[index]) == 0)
 			{
-				return index;
+				return static_cast<int>(index);
 			}
 		}
